requirements.txt
python-telegram-bot>=20.7
openai>=1.14
requests>=2.31
beautifulsoup4>=4.12
python-dotenv>=1.0
